+++ app/models/shop.rb
@@ -1,2 +1,8 @@
 class Shop < ApplicationRecord
+  has_many :business_hours
+
+  validates :name,
+            presence: true,
+            length: { minimum: 3 },
+            uniqueness: true
 end