+++ db/migrate/20220924005649_create_business_hours.rb
@@ -1,0 +1,10 @@
+class CreateBusinessHours < ActiveRecord::Migration[7.0]
+  def change
+    create_table :business_hours do |t|
+
+      t.string :opening_period
+      t.timestamps
+      t.belongs_to :shop, index: true
+    end
+  end
+end